Unhiding things that were hidden when contact details were open
A handful of icons and options were hidden presumably to save space when the contact details panel is open and squishes the main content. Now that the squishing isn't happening, there's no reason to hide these elements. This creates a more uniform user experience with not having things disappear in the normal usage of the screen.